+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.65.0] - 2023-11-15
+
+### Changed
+- Handle several stages of filters for license
+- Add security property to License
+
+
 ## [3.64.0] - 2023-11-14
 
 ### Changed